@@ -12,68 +12,68 @@
     "build:release": "web-ext build --source-dir ./build",
     "test": "jest --coverage"
   },
+  "engines": {
+    "node": ">=16.13.0"
+  },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.10.4",
-    "@percy/agent": "^0.28.2",
-    "@svgr/webpack": "^5.4.0",
-    "@types/jest": "^26.0.13",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-emotion": "^10.0.33",
-    "copy-webpack-plugin": "^6.1.0",
-    "cross-env": "^7.0.2",
-    "file-loader": "^6.1.0",
-    "html-webpack-plugin": "^4.4.1",
-    "jest": "^26.4.2",
-    "jest-localstorage-mock": "^2.4.3",
-    "prettier": "^2.1.1",
-    "ts-jest": "^26.3.0",
-    "ts-loader": "^8.0.3",
-    "typescript": "^4.0.2",
-    "web-ext": "^5.0.0",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "webpack-manifest-version-sync-plugin": "^0.0.3"
+    "@babel/core": "^7.18.6",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@emotion/babel-plugin": "^11.9.2",
+    "@svgr/webpack": "^6.2.1",
+    "@types/jest": "^28.1.4",
+    "babel-loader": "^8.2.5",
+    "copy-webpack-plugin": "^11.0.0",
+    "cross-env": "^7.0.3",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.2",
+    "jest-environment-jsdom": "^28.1.2",
+    "jest-localstorage-mock": "^2.4.21",
+    "prettier": "^2.7.1",
+    "ts-jest": "^28.0.5",
+    "ts-loader": "^9.3.1",
+    "typescript": "^4.7.4",
+    "web-ext": "^7.1.1",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-manifest-version-sync-plugin": "^0.0.4"
   },
   "dependencies": {
-    "@babel/runtime": "^7.11.2",
-    "@emotion/core": "^10.0.35",
-    "@emotion/styled": "^10.0.27",
-    "changelog-parser": "^2.8.0",
-    "classnames": "^2.2.6",
-    "core-js": "^3.6.5",
-    "emotion-theming": "^10.0.27",
-    "lodash": "^4.17.20",
-    "moment": "^2.27.0",
-    "moment-timezone": "^0.5.31",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-color": "^2.18.1",
-    "react-dom": "^16.13.1",
+    "@babel/runtime": "^7.18.6",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "changelog-parser": "^2.8.1",
+    "classnames": "^2.3.1",
+    "core-js": "^3.23.3",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.34",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-color": "^2.19.3",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-select": "^3.1.0",
-    "react-switch": "^5.0.1",
-    "react-window": "^1.8.5",
-    "typeface-inter": "^3.15.0",
-    "typeface-metropolis": "0.0.74",
-    "webextension-polyfill": "^0.6.0"
+    "react-select": "^5.4.0",
+    "react-switch": "^7.0.0",
+    "react-window": "^1.8.7",
+    "typeface-inter": "^3.18.1",
+    "typeface-metropolis": "1.1.13",
+    "webextension-polyfill": "^0.9.0"
   },
   "jest": {
     "globals": {
       "ts-jest": {
-        "tsConfig": "tsconfig.json"
+        "tsconfig": "tsconfig.json"
       }
     },
+    "testEnvironment": "jsdom",
     "setupFiles": [
       "jest-localstorage-mock"
     ],
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
     "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
     "modulePathIgnorePatterns": [
       "<rootDir>/out/"
